[fenics] move fenics bindings to bindings.fenics
