+++ src/pymordemos/parabolic_mor.py
@@ -148,9 +148,7 @@ def _discretize_fenics():
     # wrap everything as a pyMOR discretization
     ###########################################
 
-    from pymor.gui.fenics import FenicsVisualizer
-    from pymor.operators.fenics import FenicsMatrixOperator
-    from pymor.vectorarrays.fenics import FenicsVectorSpace
+    from pymor.bindings.fenics import FenicsVectorSpace, FenicsMatrixOperator, FenicsVisualizer
 
     d = InstationaryDiscretization(
         T=1.,
